requirements.txt
# streamlit
# langgraph
# langchain-core
# pydantic
# typing-extensions
# pymupdf
# # faiss-cpu
# sentence-transformers
# pinecone-client
# langchain-text-splitters
# langchain-openai
# pypdf
# python-dotenv

# streamlit
# langgraph
# langchain-core
# langchain-openai
# pinecone-client
# pydantic
# python-dotenv
# pypdf
# langchain-community
# langchain

# streamlit
# python-dotenv

# langchain-core>=0.1.48
# langchain==0.0.353
# langchain-community==0.4.1  
# langchain-text-splitters==1.0.0 

# langchain-openai
# langchain-ollama

# pinecone-client
# pypdf

# sentence-transformers
# transformers
# tensorflow
# langchain_classic

# streamlit
# langgraph
# langchain-core
# langchain-openai
# pydantic
# typing-extensions
# python-dotenv
# pinecone-client
# pypdf
# langchain
# pytesseract
# pdf2image
# pillow
# # poppler
# reportlab
# pdf2image
# openai>=1.30.0
# tiktoken>=0.7.0
# pinecone-client>=3.0.0
# numpy>=1.26.0
# PyMuPDF>=1.24.0
# pdf2image>=1.17.0
# pytesseract>=0.3.10
# Pillow>=10.0.0
# arxiv>=2.1.0
# tavily-python>=0.3.5
# requests>=2.31.0
# pydantic>=2.7.0
# langchain>=0.3.0
# langgraph
# langchain-openai
# langchain-community
# langchain-tavily
# tavily-python


# brew install poppler
# pytesseract in requirements.txt
#  brew install tesseract
#  brew install poppler
#  tesseract --version

# Core
streamlit==1.29.0
python-dotenv==1.0.0

# LangChain Core
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.3
langchain-groq==0.0.1
langgraph==0.0.20

# Vector Store
pinecone-client==3.0.0

# LLM Providers
openai==1.7.0

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10

# Document Processing
langchain-text-splitters==0.0.1

# Research Tools
arxiv==2.0.0
tavily-python==0.3.0
scholarly

# PDF Export
reportlab==4.0.7

# Utilities
pydantic==2.5.3

# Core dependencies
langchain-core
langchain-community

# Research sources
arxiv                    # arXiv API (no PDF parsing, avoiding fitz issues)
scholarly               # Free Google Scholar scraping
tavily-python           # Web search API

# Optional: For Google Scholar with SerpAPI (paid, more reliable)
# google-search-results

# Optional: If you want full PDF text extraction from arXiv
# Note: This requires PyMuPDF which caused the original error
# Only install if needed and properly configured
# pymupdf                # For PDF parsing (fitz)

# Other dependencies
requests
python-dotenv